Organize crate into modules.
